+++ lib/src/main/jni/src/utils.h
@@ -2,6 +2,6 @@
 #define utils_H__
 
 int getFileType(const char *name);
-int getFileSize(const char *path);
+long long getFileSize(const char *path);
 
 #endif // utils_H__
